requirements.txt
absl-py==0.9.0
affine==2.3.0
appdirs==1.4.3
astroid==2.4.2
atpublic==1.0
attrs==19.3.0
backcall==0.2.0
beautifulsoup4==4.9.1
black==19.10b0
bleach==3.1.5
boto3==1.13.1
botocore==1.16.1
Bottleneck==1.3.2
cachetools==4.1.1
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.2
click-plugins==1.1.1
cligj==0.5.0
cloudpickle==1.2.2
colorama==0.4.3
configobj==5.0.6
configparser==5.0.0
cycler==0.10.0
decorator==4.4.2
defusedxml==0.6.0
descartes==1.1.0
distro==1.5.0
docker-pycreds==0.4.0
docopt==0.6.2
docutils==0.15.2
dpath==2.0.1
dvc==1.1.11
earthpy==0.9.2
entrypoints==0.3
fastai==1.0.61
fastprogress==0.2.3
Fiona==1.8.13.post1
flatten-json==0.1.7
flufl.lock==3.2
funcy==1.14
future==0.18.2
geopandas==0.7.0
gitdb==4.0.4
GitPython==3.1.1
google-auth==1.19.2
google-auth-oauthlib==0.4.1
gql==0.2.0
grandalf==0.6
graphql-core==1.1
grpcio==1.30.0
humanize==2.4.0
idna==2.9
imageio==2.8.0
imgaug==0.3.0
importlib-metadata==1.6.0
inflect==3.0.2
ipykernel==5.3.0
ipython==7.15.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.17.1
Jinja2==2.11.2
jmespath==0.9.5
joblib==0.16.0
jsonpath-ng==1.5.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.3
jupyter-console==6.1.0
jupyter-core==4.6.3
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.2
mccabe==0.6.1
mistune==0.8.4
munch==2.5.0
nanotime==0.5.2
nbconvert==5.6.1
nbformat==5.0.7
networkx==2.3
notebook==6.0.3
numexpr==2.7.1
numpy==1.19.0
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opencv-python==4.3.0.36
opencv-python-headless==4.3.0.36
packaging==20.3
pandas==1.0.5
pandocfilters==1.4.2
parso==0.7.0
pathspec==0.8.0
pathtools==0.1.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
ply==3.11
prometheus-client==0.8.0
promise==2.3
prompt-toolkit==3.0.5
protobuf==3.12.2
psutil==5.7.0
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydot==1.4.1
Pygments==2.6.1
pygtrie==2.3.2
pylint==2.5.3
pyparsing==2.4.7
pyproj==2.6.1.post1
pyrsistent==0.16.0
python-dateutil==2.8.0
pytorch-ignite==0.3.0
pytorch-lightning==0.8.5
pytz==2020.1
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.1
qtconsole==4.7.5
QtPy==1.9.0
rasterio==1.1.5
regex==2020.6.8
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
s3transfer==0.3.3
scikit-image==0.17.2
scikit-learn==0.23.1
scipy==1.5.0
Send2Trash==1.5.0
sentry-sdk==0.15.1
Shapely==1.7.0
shortuuid==1.0.1
shtab==1.1.0
six==1.14.0
sklearn==0.0
skorch==0.7.0
smmap==3.0.2
snuggs==1.4.7
soupsieve==2.0.1
subprocess32==3.5.4
tabulate==0.8.7
tensorboard==2.2.2
tensorboard-plugin-wit==1.7.0
terminado==0.8.3
testpath==0.4.4
texttable==1.6.2
threadpoolctl==2.1.0
tifffile==2020.6.3
toml==0.10.1
torch==1.5.1
torchbearer==0.5.1
torchvision==0.6.1
tornado==6.0.4
tqdm==4.46.0
traitlets==4.3.3
treelib==1.6.1
typed-ast==1.4.1
typeguard==2.9.1
urllib3==1.25.9
voluptuous==0.11.7
wandb==0.9.1
watchdog==0.10.3
wcwidth==0.2.4
webencodings==0.5.1
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
zc.lockfile==2.0
zipp==3.1.0
